Renamed sample
