+++ data/shaders/glsl/gltfscenerendering/scene.frag.spv
