+++ data/shaders/glsl/gltfscenerendering/scene.vert
