+++ data/shaders/hlsl/gltfscenerendering/scene.frag.spv
