+++ data/shaders/hlsl/gltfscenerendering/scene.vert.spv
